+++ sdk/spring/spring-messaging-azure-storage-queue/src/test/java/com/azure/spring/storage/queue/core/StorageQueueTemplateSendTests.java
@@ -3,16 +3,21 @@
 
 package com.azure.spring.storage.queue.core;
 
-import com.azure.spring.messaging.core.SendOperationTest;
 import com.azure.spring.storage.queue.core.factory.StorageQueueClientFactory;
 import com.azure.storage.queue.QueueAsyncClient;
 import com.azure.storage.queue.models.SendMessageResult;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.support.MessageBuilder;
 import reactor.core.publisher.Mono;
 
+import static org.assertj.core.api.Fail.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -21,7 +26,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class StorageQueueTemplateSendTests extends SendOperationTest<StorageQueueTemplate> {
+public class StorageQueueTemplateSendTests {
 
     @Mock
     private StorageQueueClientFactory mockClientFactory;
@@ -30,6 +35,9 @@ public class StorageQueueTemplateSendTests extends SendOperationTest<StorageQueu
     private QueueAsyncClient mockClient;
 
     private AutoCloseable closeable;
+    protected String destination = "storage-queue";
+    protected Message<?> message = MessageBuilder.withPayload("test").build();
+    protected StorageQueueTemplate storageQueueTemplate;
 
     @BeforeEach
     public void init() {
@@ -46,21 +54,40 @@ public class StorageQueueTemplateSendTests extends SendOperationTest<StorageQueu
         when(this.mockClientFactory.createQueueClient(eq(destination))).thenReturn(mockClient);
         when(this.mockClient.sendMessage(anyString())).thenReturn(Mono.just(new SendMessageResult()));
 
-        this.sendOperation = new StorageQueueTemplate(mockClientFactory);
+        this.storageQueueTemplate = new StorageQueueTemplate(mockClientFactory);
     }
 
-    @Override
+    @Test
+    public void testSend() {
+        final Mono<Void> mono = this.storageQueueTemplate.sendAsync(destination, message);
+
+        assertNull(mono.block());
+        verifySendCalled(1);
+    }
+
+    @Test
+    public void testSendFailure() {
+        String errorMessage = "Send failed.";
+        setupError(errorMessage);
+        Mono<Void> mono = this.storageQueueTemplate.sendAsync(destination, this.message);
+
+        try {
+            mono.block();
+            fail("Test should fail.");
+        } catch (Exception e) {
+            assertEquals(errorMessage, e.getMessage());
+        }
+    }
+
     protected void setupError(String errorMessage) {
         when(this.mockClient.sendMessage(any(String.class)))
             .thenReturn(Mono.error(new IllegalArgumentException(errorMessage)));
     }
 
-    @Override
     protected void verifySendCalled(int times) {
         verify(this.mockClient, times(times)).sendMessage(isA(String.class));
     }
 
-    @Override
     protected void verifyGetClientCreator(int times) {
         verify(this.mockClientFactory, times(times)).createQueueClient(this.destination);
     }